asasas

+++ internal/dto/content_dto.go
@@ -12,7 +12,7 @@ type GetDetailContentResDTO struct {
 	ContentID uint                        `json:"content_id"`
 	Title     string                      `json:"title"`
 	Type      string                      `json:"type"`
-	Duration  int                         `json:"duration"`
+	Duration  uint16                      `json:"duration"`
 	URL       string                      `json:"url"`
 	Airport   GetSummaryAirportResDTO     `json:"airport"`
 	Playlists []*GetSummaryPlaylistResDTO `json:"playlists"`
